Fixed preprocessor inheritance to not monkey patch base methods
